pyxis: Force build scudo free 32 bit variant of libc
After playing with gcam port, stream configurations for certain
modes I found that camera blobs are not compatible with scudo and
such usage leads to crashes of camera provider like [1].
Switch to scudo-free libc variant resolves this issues. As the
camera blobs are 32bit it is not needed to switch to scudo-free
libc system wide.

[1] https://pastebin.com/W5shnvEM

Change-Id: I5465b79b8c009550cd8a78666c86c5b4d0b242ae